backend/pyproject.toml
[project]
name = "mlb-notification-backend"
version = "0.1.0"
description = "MLB Japanese player notification backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy==2.0.36",
    "aiosqlite==0.20.0",
    "redis==5.2.0",
    "apscheduler==3.10.4",
    "httpx==0.27.2",
    "pydantic-settings==2.6.0",
    "pytz==2024.2",
    "greenlet>=3.3.2",
]

[dependency-groups]
dev = []
